assignment-6/requirements.txt
dotenv==0.9.9
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tabulate==0.9.0
tzdata==2025.1
